Exciting new blast to XMFAp tooling
This tool takes blast results against a small group of genomes,
a list of genomes and gene calls that went into that blast database,
and then can output an XMFA file representing the alignments.

It does this via figuring out blast clusters and running those through
clustal for further alignments.

The resulting files should work completely with the rest of our tooling,
taking advantage of some of the awesome things we can do with XMFA