Fixed re-open tool.
CC @elenimijalis. The reason things were showing up in Apollo as partial
exons/CDSs is because Apollo was rejecting the translation. After much
debugging it turned out that after re-opening, the sequence of a given
feature changed. (I added a print statement before/after re-open process
and printed out the sequence of a single feature.) Apparently slicing
the record object applies to the sequence as well. We were
double-cutting and this caused major issues.

Will fix + push + run the create_annotation script.