Sort on general (record IDX) + specific (feat loc)
If we sort only on record location, then we hit an issue on the assert
statement, the featuers are not ordered consistently between sort
invocations (good). So we have to add a "general" component which is the
record and overriding location, but then a specific to ensure consistent
sorting in-between features in the same genome.