Fix translation bug: accidental stop codon at end
When a TranslationError occured we were not translating with cds=True,
we used cds=False which includes stop codons. This resulted in one case
where the very last gene (which had an invalid start codon) having a
stop codon at the end of the sequence. We now actually trim that
sequence instead of leaving it with a stop