Fix toxic bug in rename tool
cc @moffmade When the entries were read in from the file, they contained
a newline at the end. This was propagated into Apollo. Whenever we tried
to export features from that apollo record, we found that export broke,
because it tries to read GFF3 and that stray newline breaks everything,
preventing workflows from being run on that data.