Update gff3_rebase.xml
Made edits to text.  There is a display error when trying to show the sample gff3 output shown in the xml in the Galaxy tool window, it seems to not be reading the text and giving errors.  What are the field codes for having text appear verbatim, preferably in a different (monospace) font or blocked out?